+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 0.20.0 (TBD)
 
+#### Enhancements
+
+#### Changes
+- Added missing implementations of `proptest::Arbitrary` for non-`BasicBlockNode` variants of `MastNode` ([#2335](https://github.com/0xMiden/miden-vm/pull/2335)).
+
 ## 0.19.0 (2025-11-1)
 
 #### Enhancements